refactored hiders into separate package
