+++ src/main/java/pl/betoncraft/betonquest/compatibility/citizens/CitizensParticle.java
@@ -11,7 +11,7 @@ import org.bukkit.entity.Player;
 import org.bukkit.scheduler.BukkitRunnable;
 import pl.betoncraft.betonquest.BetonQuest;
 import pl.betoncraft.betonquest.compatibility.effectlib.EffectLibIntegrator;
-import pl.betoncraft.betonquest.compatibility.protocollib.NPCHider;
+import pl.betoncraft.betonquest.compatibility.protocollib.hider.NPCHider;
 import pl.betoncraft.betonquest.config.Config;
 import pl.betoncraft.betonquest.config.ConfigPackage;
 import pl.betoncraft.betonquest.exceptions.ObjectNotFoundException;
